Separate upload-from-file from dataset upload panel
