+++ asreview/webapp/src/PreReviewComponents/index.js
@@ -8,6 +8,7 @@ export { default as SearchResult } from '../PreReviewComponents/SearchResult'
 export { default as PaperCard } from '../PreReviewComponents/PaperCard'
 export { default as ProjectInit } from '../PreReviewComponents/ProjectInit'
 export { default as ProjectUpload } from '../PreReviewComponents/ProjectUpload'
+export { default as ProjectUploadFile } from '../PreReviewComponents/ProjectUploadFile'
 export { default as ProjectUploadBenchmarkDatasets } from '../PreReviewComponents/ProjectUploadBenchmarkDatasets'
 export { default as ProjectUploadPluginDatasets } from '../PreReviewComponents/ProjectUploadPluginDatasets'
 export { default as BenchmarkDataset } from '../PreReviewComponents/BenchmarkDataset'
